requirements.txt
tensorflow>=2.8.0
numpy>=1.20.0
matplotlib>=3.4.0
streamlit>=1.10.0
streamlit-drawable-canvas>=0.9.0
pillow>=9.0.0